refactor testComputeHelper
